refactor(memberfarepolicy): 패키지명 수정
