+++ src/main/java/nextstep/subway/path/domain/Path.java
@@ -12,7 +12,7 @@ import nextstep.subway.error.ErrorMessage;
 import nextstep.subway.line.domain.Distance;
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.Section;
-import nextstep.subway.path.farePolicy.MemberDiscountPolicy;
+import nextstep.subway.path.memberfarepolicy.MemberDiscountPolicy;
 import nextstep.subway.path.dto.PathResponse;
 import nextstep.subway.station.domain.Station;
